backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
google-genai
httpx
pydantic
python-multipart
websockets
